refactor: split WorkHoursPage into per-view components and API hook
Break the 2642-line WorkHoursPage.tsx into focused modules:
- views/{DayView,WeekView,MonthView,SettingsView}.tsx
- components/{FlexTrendChart,EmojiPickerDropdown}.tsx
- useWorkHoursApi.ts — typed wrappers for every /api/workhours/* call
  and the workhours-related /api/settings/preferences calls
- types.ts — shared interfaces (WorkDay, DaySummary, LeaveBalance, ...)
- dateUtils.ts — pure date/format/holiday helpers
- presetIcons.ts — deduction-preset icon helpers

WorkHoursPage now only holds the view switcher and shared
ViewMode/currentDate state. No behavior, layout, or endpoint changes;
the live punch-in continues to survive view switches via server restore.

Bead: Hytte-ieus
